+++ src/app/auth/hooks/index.ts
@@ -1,0 +1,7 @@
+import trackUser from './trackUser';
+import untrackUser from './untrackUser';
+
+export {
+  trackUser,
+  untrackUser,
+};
