+++ src/app/auth/hooks/trackUser.ts
@@ -1,9 +1,9 @@
-import googleAnalyticsClient from '../../gateways/googleAnalyticsClient';
-import { receiveUser } from '../auth/actions';
-import { isAcceptCookiesNeeded } from '../notifications/acceptCookies';
-import { acceptCookies } from '../notifications/actions';
-import { ActionHookBody } from '../types';
-import { actionHook } from '../utils';
+import googleAnalyticsClient from '../../../gateways/googleAnalyticsClient';
+import { isAcceptCookiesNeeded } from '../../notifications/acceptCookies';
+import { acceptCookies } from '../../notifications/actions';
+import { ActionHookBody } from '../../types';
+import { actionHook } from '../../utils';
+import { receiveUser } from '../actions';
 
 export const trackUserHookBody: ActionHookBody<typeof receiveUser> = (middleware) => async({payload}) => {
   if (payload.isNotGdprLocation) {
