+++ src/app/auth/index.ts
@@ -1,4 +1,5 @@
 import * as actions from './actions';
+import * as hooks from './hooks';
 import * as init from './init';
 import reducer from './reducer';
 import * as types from './types';
@@ -8,4 +9,5 @@ export {
   types,
   reducer,
   actions,
+  hooks,
 };
