+++ src/app/hooks/index.ts
@@ -1,9 +1,5 @@
 import trackPageView from './trackPageView';
-import trackUser from './trackUser';
-import untrackUser from './untrackUser';
 
 export {
   trackPageView,
-  trackUser,
-  untrackUser,
 };
