+++ src/app/index.tsx
@@ -49,6 +49,7 @@ const hooks = [
   ...content.hooks,
   ...Object.values(head.hooks),
   ...Object.values(notifications.hooks),
+  ...Object.values(auth.hooks),
   ...Object.values(appHooks),
 ];
